# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vokal Director API",
    "description": "Контракт между фронтендом и сервером Vokal Director.\n\n**Что здесь важно знать до чтения адресов.**\n\n* Большая часть адресов отвечает `501`. Это не заготовка на будущее: схема\n  запроса и ответа настоящая и полная, реализации за адресом еще нет. В\n  `details.missing` сказано, чего именно не хватает. Заглушек, возвращающих\n  правдоподобные данные, в API нет — мок в ответе сервера хуже отказа.\n* Ошибка всегда выглядит одинаково: `{\"error\": {\"code\", \"message\",\n  \"requestId\", \"details\"}}`.\n* Имена полей — camelCase. Идентификаторы — непрозрачные строки.\n* Незнакомое поле в теле запроса — ошибка `422`, а не тихо отброшенное\n  значение.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/versions": {
      "get": {
        "tags": [
          "версии"
        ],
        "summary": "Версии аранжировки",
        "description": "От исходной к последней, вместе с родителями и перечнем изменений.",
        "operationId": "list_versions_api_projects__project_id__versions_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/versions/{version_id}": {
      "get": {
        "tags": [
          "версии"
        ],
        "summary": "Одна версия",
        "operationId": "read_version_api_projects__project_id__versions__version_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор версии",
              "title": "Version Id"
            },
            "description": "Идентификатор версии"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArrangementVersionOut"
                }
              }
            }
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/versions/{version_id}/select": {
      "patch": {
        "tags": [
          "версии"
        ],
        "summary": "Сделать версию текущей",
        "description": "Переключение без создания новой версии: история при этом не меняется. Если материалы версии сейчас в работе у более поздней, переключение отвечает `409` и указывает на откат — он остается в истории.",
        "operationId": "select_version_api_projects__project_id__versions__version_id__select_patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор версии",
              "title": "Version Id"
            },
            "description": "Идентификатор версии"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionSelectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArrangementVersionOut"
                }
              }
            }
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Состояние объекта изменилось: обновите данные и повторите.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Данные запроса не прошли проверку схемы.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/versions/{version_id}/rollback": {
      "post": {
        "tags": [
          "версии"
        ],
        "summary": "Откатиться на указанную версию",
        "description": "Цель — любая версия проекта, глубина не ограничена. Создается новая версия со снимком материалов целевой; материалы, снимка которых не было, помечаются к пересборке.",
        "operationId": "rollback_version_api_projects__project_id__versions__version_id__rollback_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор версии",
              "title": "Version Id"
            },
            "description": "Идентификатор версии"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionRollbackRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionRollbackResponse"
                }
              }
            }
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Состояние объекта изменилось: обновите данные и повторите.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Данные запроса не прошли проверку схемы.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArrangementVersionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "kind": {
            "$ref": "#/components/schemas/VersionKind"
          },
          "parentVersionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentversionid",
            "description": "Родительская версия. Пусто у исходной."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "createdBy": {
            "type": "string",
            "title": "Createdby",
            "description": "Кто создал: человек или действие директора"
          },
          "status": {
            "$ref": "#/components/schemas/VersionStatus"
          },
          "changes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Changes",
            "description": "Что изменилось относительно родителя"
          },
          "artifactsSnapshot": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ArtifactOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifactssnapshot",
            "description": "Материалы, с которыми создана версия. По ним делается откат. Пусто — материалы этой версии неизвестны, и откат пометит перенесенные к пересборке."
          },
          "isCurrent": {
            "type": "boolean",
            "title": "Iscurrent",
            "description": "Выбрана ли версия сейчас"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "label",
          "kind",
          "createdAt",
          "createdBy",
          "status",
          "changes",
          "isCurrent"
        ],
        "title": "ArrangementVersionOut",
        "description": "Версия аранжировки."
      },
      "ArtifactAudience": {
        "type": "string",
        "enum": [
          "all",
          "band",
          "teacher",
          "student",
          "instrument"
        ],
        "title": "ArtifactAudience",
        "description": "Кому предназначен материал. Разным ролям выдается разное."
      },
...
      "ArtifactFormat": {
        "type": "string",
        "enum": [
          "PDF",
          "MIDI",
          "WAV",
          "ZIP",
          "MusicXML",
          "VIEW"
        ],
        "title": "ArtifactFormat"
      },
...
      "ArtifactOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/ArtifactType"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "format": {
            "$ref": "#/components/schemas/ArtifactFormat"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/ArtifactStatus"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Уверенность: 0..1, показывается процентом"
          },
          "audience": {
            "$ref": "#/components/schemas/ArtifactAudience"
          },
          "isStale": {
            "type": "boolean",
            "title": "Isstale",
            "description": "Материал не соответствует текущей версии аранжировки"
          },
          "preview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArtifactPreviewOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "name",
          "format",
          "description",
          "status",
          "confidence",
          "audience",
          "isStale"
        ],
        "title": "ArtifactOut",
        "description": "Материал Stage Pack."
      },
      "ArtifactPreviewKind": {
        "type": "string",
        "enum": [
          "notation",
          "waveform",
          "midi",
          "bundle",
          "text"
        ],
        "title": "ArtifactPreviewKind"
      },
      "ArtifactPreviewOut": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ArtifactPreviewKind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "lines": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "title",
          "lines"
        ],
        "title": "ArtifactPreviewOut",
        "description": "Предпросмотр без скачивания файла."
      },
      "ArtifactStatus": {
        "type": "string",
        "enum": [
          "ready",
          "draft",
          "needs_review",
          "rebuild_required",
          "pending"
        ],
        "title": "ArtifactStatus"
      },
      "ArtifactType": {
        "type": "string",
        "enum": [
          "score",
          "part",
          "tab",
          "chords",
          "lyrics",
          "midi",
          "stem",
          "minus",
          "click",
          "practice",
          "teacher",
          "student",
          "zip"
        ],
        "title": "ArtifactType"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Машинный код для ветвления на клиенте"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Текст для показа пользователю"
          },
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestid",
            "description": "Идентификатор запроса: по нему поддержка ищет след в логах"
          },
          "details": {
            "title": "Details",
            "description": "Подробности, зависят от кода. Для validation_error — {fields: [{field, reason}]}, для not_implemented — {endpoint, missing, docs}."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "description": "Тело ошибки. Форма задана `app/core/errors.py` и одинакова для всего API."
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "Конверт ошибки: любой ответ с кодом 4xx/5xx выглядит так."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VersionKind": {
        "type": "string",
        "enum": [
          "original",
          "band",
          "easy",
          "original-like",
          "advanced",
          "concert",
          "ensemble",
          "after-rehearsal",
          "after-lesson"
        ],
        "title": "VersionKind"
      },
      "VersionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ArrangementVersionOut"
            },
            "type": "array",
            "title": "Items"
          },
          "currentVersionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentversionid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "VersionListResponse"
      },
      "VersionRollbackRequest": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Label",
            "description": "Подпись новой версии. Пусто — соберется из подписи целевой версии."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "VersionRollbackRequest",
        "description": "Откат на версию, указанную в адресе."
      },
      "VersionRollbackResponse": {
        "properties": {
          "version": {
            "$ref": "#/components/schemas/ArrangementVersionOut"
          },
          "restoredFromVersionId": {
            "type": "string",
            "title": "Restoredfromversionid"
          },
          "restoredArtifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactOut"
            },
            "type": "array",
            "title": "Restoredartifacts"
          },
          "staleArtifactIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Staleartifactids",
            "description": "Материалы, которые придется пересобрать: снимка для них не было"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "restoredFromVersionId",
          "restoredArtifacts"
        ],
        "title": "VersionRollbackResponse",
        "description": "Результат отката.\n\nНовая версия, а не переписанная история: откат — такое же событие, как\nправка, и он обязан остаться в списке версий. Иначе по истории нельзя\nпонять, почему материалы вдруг стали прежними."
      },
      "VersionSelectRequest": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "VersionSelectRequest",
        "description": "Переключение текущей версии без создания новой."
      },
      "VersionStatus": {
        "type": "string",
        "enum": [
          "draft",
          "needs_review",
          "approved",
          "distributed"
        ],
        "title": "VersionStatus"
      }
    }
  }
}